requirements.txt
asgiref==3.8.1
Django==5.1.2
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
phonenumbers==9.0.3
pillow==11.1.0
python-dotenv==1.1.0
sqlparse==0.5.3
